refactor: ViewHolderのサイズが大きくなってきて視認性が悪いので、別ファイルに切り出した
