+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/timeline/TimelineListAdapter.kt
@@ -3,14 +3,9 @@ package net.pantasystem.milktea.note.timeline
 
 import android.content.Context
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.core.view.isVisible
 import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.flowWithLifecycle
-import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
@@ -19,27 +14,17 @@ import com.bumptech.glide.ListPreloader.PreloadModelProvider
 import com.bumptech.glide.RequestBuilder
 import com.bumptech.glide.integration.recyclerview.RecyclerViewPreloader
 import com.bumptech.glide.util.FixedPreloadSizeProvider
-import com.google.android.flexbox.FlexboxLayout
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
 import net.pantasystem.milktea.common.glide.GlideApp
 import net.pantasystem.milktea.common.glide.GlideUtils
-import net.pantasystem.milktea.common_android.ui.FontSizeHelper.specialPointToPixel
 import net.pantasystem.milktea.model.setting.DefaultConfig
 import net.pantasystem.milktea.model.setting.LocalConfigRepository
 import net.pantasystem.milktea.note.R
 import net.pantasystem.milktea.note.databinding.ItemHasReplyToNoteBinding
 import net.pantasystem.milktea.note.databinding.ItemNoteBinding
-import net.pantasystem.milktea.note.databinding.ItemTimelineEmptyBinding
-import net.pantasystem.milktea.note.databinding.ItemTimelineErrorBinding
-import net.pantasystem.milktea.note.reaction.ReactionViewData
 import net.pantasystem.milktea.note.timeline.viewmodel.TimelineListItem
 import net.pantasystem.milktea.note.view.NoteCardAction
 import net.pantasystem.milktea.note.view.NoteCardActionListenerAdapter
-import net.pantasystem.milktea.note.view.NoteUserRoleBadgeBinder.setUserRoleBadge
 import net.pantasystem.milktea.note.viewmodel.HasReplyToNoteViewData
-import net.pantasystem.milktea.note.viewmodel.PlaneNoteViewData
 import java.util.Collections
 
 class TimelineListAdapter(
@@ -48,7 +33,7 @@ class TimelineListAdapter(
     val onRefreshAction: () -> Unit,
     val onReauthenticateAction: () -> Unit,
     val onAction: (NoteCardAction) -> Unit,
-) : ListAdapter<TimelineListItem, TimelineListAdapter.TimelineListItemViewHolderBase>(
+) : ListAdapter<TimelineListItem, TimelineListItemViewHolderBase>(
     object : DiffUtil.ItemCallback<TimelineListItem>() {
         override fun areContentsTheSame(
             oldItem: TimelineListItem,
@@ -83,146 +68,10 @@ class TimelineListAdapter(
     private val urlPreviewListRecyclerViewPool = RecyclerView.RecycledViewPool()
     private val manyFilePreviewListViewRecyclerViewPool = RecyclerView.RecycledViewPool()
 
-    sealed class TimelineListItemViewHolderBase(view: View) : RecyclerView.ViewHolder(view)
-
-    sealed class NoteViewHolderBase<out T : ViewDataBinding>(
-        view: View,
-        val reactionCountBinder: ReactionCountItemsFlexboxLayoutBinder
-    ) :
-        TimelineListItemViewHolderBase(view) {
-
-        abstract val binding: T
-        abstract val lifecycleOwner: LifecycleOwner
-//        abstract val reactionCountsView: RecyclerView
-        abstract val flexboxLayout: FlexboxLayout
-        abstract val noteCardActionListenerAdapter: NoteCardActionListenerAdapter
-
-//        private var reactionCountAdapter: ReactionCountAdapter? = null
-
-
-        abstract fun onBind(note: PlaneNoteViewData)
-
-        private var mCurrentNote: PlaneNoteViewData? = null
-
-        fun bind(note: PlaneNoteViewData) {
-
-            unbind()
-
-            mCurrentNote = note
-            bindReactionCounter()
-
-            onBind(mCurrentNote!!)
-            binding.lifecycleOwner = lifecycleOwner
-            binding.executePendingBindings()
-        }
-
-        private var job: Job? = null
-
-        fun unbind() {
-            job?.cancel()
-//            reactionCountsView.itemAnimator?.endAnimations()
-
-            mCurrentNote = null
-        }
-
-        private fun bindReactionCounter() {
-            val note = mCurrentNote!!
-            job = note.reactionCountsViewData.onEach { counts ->
-                bindReactionCountVisibility(counts)
-            }.flowWithLifecycle(lifecycleOwner.lifecycle).launchIn(lifecycleOwner.lifecycleScope)
-        }
-
-        private fun bindReactionCountVisibility(reactionCounts: List<ReactionViewData>?) {
-            reactionCountBinder.bindReactionCounts(
-                flexboxLayout,
-                mCurrentNote,
-                reactionCounts ?: emptyList(),
-            )
-
-            flexboxLayout.visibility = if (reactionCounts.isNullOrEmpty()) {
-                View.GONE
-            } else {
-                View.VISIBLE
-            }
-        }
-
-    }
-
-    class LoadingViewHolder(view: View) : TimelineListItemViewHolderBase(view)
-
-    class ErrorViewHolder(val binding: ItemTimelineErrorBinding) :
-        TimelineListItemViewHolderBase(binding.root) {
-        fun bind(item: TimelineListItem.Error) {
-            binding.errorItem = item
-            binding.errorView.isVisible = false
-            binding.showErrorMessageButton.isVisible = true
-            binding.errorView.text = item.throwable.toString()
-            binding.showErrorMessageButton.setOnClickListener {
-                binding.errorView.isVisible = true
-            }
-        }
-    }
-
-    class EmptyViewHolder(val binding: ItemTimelineEmptyBinding) :
-        TimelineListItemViewHolderBase(binding.root)
-
     enum class ViewHolderType {
         NormalNote, HasReplyToNote, Loading, Empty, Error
     }
 
-    inner class NoteViewHolder(override val binding: ItemNoteBinding, binder: ReactionCountItemsFlexboxLayoutBinder) :
-        NoteViewHolderBase<ItemNoteBinding>(binding.root, binder) {
-
-        override val lifecycleOwner: LifecycleOwner
-            get() = this@TimelineListAdapter.lifecycleOwner
-//        override val reactionCountsView: RecyclerView
-//            get() = binding.simpleNote.reactionView
-
-
-        override val flexboxLayout: FlexboxLayout
-            get() = binding.simpleNote.reactionView
-
-        override val noteCardActionListenerAdapter: NoteCardActionListenerAdapter
-            get() = this@TimelineListAdapter.cardActionListener
-
-        override fun onBind(note: PlaneNoteViewData) {
-            binding.note = note
-            binding.noteCardActionListener = noteCardActionListenerAdapter
-            val badgeIconSize = binding.root.context.specialPointToPixel(note.config.value.noteHeaderFontSize).toInt()
-            binding.simpleNote.badgeRoles.apply {
-                setUserRoleBadge(binding.simpleNote.noteLayout, note.toShowNote.user.iconBadgeRoles, badgeIconSize)
-            }
-        }
-
-
-    }
-
-    inner class HasReplyToNoteViewHolder(override val binding: ItemHasReplyToNoteBinding, binder: ReactionCountItemsFlexboxLayoutBinder) :
-        NoteViewHolderBase<ItemHasReplyToNoteBinding>(binding.root, binder) {
-        override val lifecycleOwner: LifecycleOwner
-            get() = this@TimelineListAdapter.lifecycleOwner
-
-        override val flexboxLayout: FlexboxLayout
-            get() = binding.simpleNote.reactionView
-
-//        override val reactionCountsView: RecyclerView
-//            get() = binding.simpleNote.reactionView
-
-        override val noteCardActionListenerAdapter: NoteCardActionListenerAdapter
-            get() = this@TimelineListAdapter.cardActionListener
-
-        override fun onBind(note: PlaneNoteViewData) {
-            if (note is HasReplyToNoteViewData) {
-                binding.hasReplyToNote = note
-                binding.noteCardActionListener = noteCardActionListenerAdapter
-                binding.simpleNote.badgeRoles.apply {
-                    val badgeIconSize = context.specialPointToPixel(note.config.value.noteHeaderFontSize).toInt()
-                    setUserRoleBadge(binding.simpleNote.noteLayout, note.toShowNote.user.iconBadgeRoles, badgeIconSize)
-                }
-            }
-        }
-
-    }
 
 
     override fun getItemViewType(position: Int): Int {
@@ -300,7 +149,7 @@ class TimelineListAdapter(
                     headerFontSize = config.noteHeaderFontSize,
                     contentFontSize = config.noteContentFontSize,
                 )
-                NoteViewHolder(binding, reactionCountItemsFlexboxLayoutBinder)
+                NoteViewHolder(binding, reactionCountItemsFlexboxLayoutBinder, lifecycleOwner, cardActionListener)
             }
 
             ViewHolderType.HasReplyToNote -> {
@@ -319,7 +168,7 @@ class TimelineListAdapter(
                     headerFontSize = config.noteHeaderFontSize,
                     contentFontSize = config.noteContentFontSize,
                 )
-                HasReplyToNoteViewHolder(binding, reactionCountItemsFlexboxLayoutBinder)
+                HasReplyToNoteViewHolder(binding, reactionCountItemsFlexboxLayoutBinder, lifecycleOwner, cardActionListener)
             }
 
             ViewHolderType.Loading -> {
